+++ core-bundle.ps1
@@ -1,13 +1,18 @@
 
 function Task-List {
-    $config.bundles.Keys | Sort-Object | ForEach-Object {
+    Param(
+        [string] $bundle
+    )
+
+    $bundles = if ($bundle) { @($bundle) } else { $config.bundles.Keys }
+
+    $bundles | Sort-Object | ForEach-Object {
         $bundleName = "[$_]".PadRight(15)
         
         $config.bundles[$_] | Sort-Object -Property Alias | ForEach-Object {
             $task = $_
-            $bundle = 
-            $description = PwshRun-ExpandVariables $task.Description $config.vars
-            $example = PwshRun-ExpandVariables $task.Example $config.vars
+            $description = PwshRun-ExpandVariables $task.Description
+            $example = PwshRun-ExpandVariables $task.Example
             Write-Output "$bundleName $($task.Alias) - $description"
             Write-Output "                  > $example"
         }
@@ -16,6 +21,14 @@ function Task-List {
 
 function Task-Metadata {
     $config.tasks
+}
+
+function Task-ShowVars {
+    $config.vars
+}
+
+function Task-ShowSettings {
+    $config.settings
 }
 
 PwshRun-RegisterTasks "core" @(
@@ -30,5 +43,17 @@ PwshRun-RegisterTasks "core" @(
         Command = "Task-Metadata";
         Description = "Get metadata of all available tasks";
         Example = "`$RUNNER task:metadata";
+    },
+    @{
+        Alias = "task:vars";
+        Command = "Task-ShowVars";
+        Description = "Show all task variables";
+        Example = "`$RUNNER task:vars";
+    },
+    @{
+        Alias = "task:settings";
+        Command = "Task-ShowSettings";
+        Description = "Show all task settings";
+        Example = "`$RUNNER task:settings";
     }
 )